debugged learned rna, still not sampling
