+++ graphlearn/abstract_graphs/learned_RNA.py
@@ -1,20 +1,24 @@
 import RNA as rna
-from graphlearn.abstract_graphs.abstract import AbstractWrapper
 from graphlearn.processing import PreProcessor
+from graphlearn.abstract_graphs.learned_molecules import PreProcessor as default_preprocessor
 from graphlearn.utils import draw
 
 
 class RnaPreProcessor(PreProcessor):
 
-    def __init__(self,base_thickness_list=[2], kmeans_clusters=2):
+    def __init__(self,base_thickness_list=[2], kmeans_clusters=2,structure_mod=True):
         self.base_thickness_list= base_thickness_list
         self.kmeans_clusters=kmeans_clusters
+        self.structure_mod=structure_mod
 
     def fit(self, inputs):
-        self.NNmodel=rna.NearestNeighborFolding()
-        self.NNmodel.fit(inputs,4)
-        abstr_input = [self._sequence_to_base_graph(seq) for seq in inputs ]
-        self.make_abstract = PreProcessor(self.base_thickness_list,self.kmeans_clusters)
+
+        self.NNmodel=rna.NearestNeighborFolding(4)
+        self.NNmodel.fit(inputs)
+
+        abstr_input = [ self._sequence_to_base_graph(seq) for seq in inputs ]
+        self.make_abstract = default_preprocessor(self.base_thickness_list,self.kmeans_clusters)
+        self.make_abstract.set_param(self.vectorizer)
         self.make_abstract.fit(abstr_input)
 
         return self
@@ -29,9 +33,9 @@ class RnaPreProcessor(PreProcessor):
         -------
         graphwrapper iterator
         '''
-        inputs=list(inputs)
 
-        self.fit(inputs,self.vectorizer)
+        inputs=list(inputs)
+        self.fit(inputs)
         return self.transform(inputs)
 
     def re_transform_single(self, graph):
@@ -58,9 +62,11 @@ class RnaPreProcessor(PreProcessor):
 
 
 
-    def _sequence_to_base_graph(self,sequence):
+    def _sequence_to_base_graph(self, sequence):
+
         structure = self.NNmodel.transform_single(sequence)
-        structure,sequence= rna.fix_structure(structure,sequence)
+        if self.structure_mod:
+            structure,sequence= rna.fix_structure(structure,sequence)
         base_graph = rna.converter.sequence_dotbracket_to_graph(seq_info=sequence, \
                                                                 seq_struct=structure)
         return base_graph
@@ -81,7 +87,10 @@ class RnaPreProcessor(PreProcessor):
         for sequence in sequences:
             if type(sequence)==str:
                 structure = self.NNmodel.transform_single(sequence)
-                structure,sequence= rna.fix_structure(structure,sequence)
+
+                if self.structure_mod:
+                    structure,sequence= rna.fix_structure(structure,sequence)
+
                 base_graph = rna.converter.sequence_dotbracket_to_graph(seq_info=sequence, \
                                                                         seq_struct=structure)
                 abstract_graph=self.make_abstract.abstract(base_graph.copy())
@@ -100,19 +109,3 @@ class RnaPreProcessor(PreProcessor):
         return result
 
 
-"""
-below: mole version
-"""
-
-
-class ScoreGraphWrapper(AbstractWrapper):
-    def abstract_graph(self):
-        return self._abstract_graph
-
-    #def __init__(self,graph,vectorizer=eden.graph.Vectorizer(), base_thickness_list=None):
-    def __init__(self, abstr,graph,vectorizer=None, base_thickness_list=None):
-        self.some_thickness_list=base_thickness_list
-        self.vectorizer=vectorizer
-        self._base_graph=graph
-        self._abstract_graph=abstr
-        self._mod_dict={}
